requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudinary==1.44.1
cryptography==45.0.5
Django==5.2.4
djangorestframework==3.16.0
fastapi==0.115.14
h11==0.16.0
idna==3.10
lxml==6.0.0
numpy==2.3.1
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyMuPDF==1.26.3
PyPDF2==3.0.1
pypdfium2==4.30.0
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
selenium==4.34.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
starlette==0.46.2
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websocket-client==1.8.0
wsproto==1.2.0
